breaking down ragfilemodule
